1) If the response if of type None, just accept anything and do not validate.
As per: wordnik's swagger spec, operation type for pet.addPet is void
but still the response sent back is {u'code': 200, u'message':
u'SUCCESS'}, which has no concrete structure to check against.

2) Accept Py model in request body.

Swagger-py lets you build any instance of a model without any check.
But, when it is passed as a request body param, validation checks kick
in. _flat_dict() is used to recursively convert the model into a dict
which json can dump later.

While this conversion, type is checked against the swagger spec.
If any of the attrs are None (that happens if a default instance is
created and any of the attrs are complex) they are removed from the
dict. And if they were required for that model, Error is raised.